js and php was moved to external files in user.php
